gradio_ui/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
fastapi==0.87.0
ffmpy==0.3.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
gradio==3.11.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.1
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
linkify-it-py==1.0.3
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.6.2
mdit-py-plugins==0.3.1
mdurl==0.1.2
multidict==6.0.2
numpy==1.23.5
orjson==3.8.2
packaging==21.3
pandas==1.5.2
paramiko==2.12.0
Pillow==9.3.0
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
pydub==0.25.1
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.6
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
typing_extensions==4.4.0
uc-micro-py==1.0.1
urllib3==1.26.13
uvicorn==0.20.0
websockets==10.4
yarl==1.8.1
